Delete any TLP keywords in the codes.
TLP is similar to ILP. It should be measured at runtime to identify how parallel it is at that time.
However, here we are only calculating how many threads are in running or runnable state, which is a clue for us to know
how many computation power (cores/cpu) is needed.